% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_converter.R
\name{K_to_P}
\alias{K_to_P}
\title{Kilogram conversion}
\usage{
K_to_P(K_mass)
}
\arguments{
\item{K_mass}{The mass in Kilograms}
}
\value{
The mass in Pounds
}
\description{
Convert mass in Kilogram to Pounds
}
\examples{
mass1 <- K_to_P(69);
mass2 <- K_to_P( c(55, 69, 71) );
}
