% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_converter.R
\name{P_to_K}
\alias{P_to_K}
\title{Pounds conversion}
\usage{
P_to_K(P_mass)
}
\arguments{
\item{P_mass}{The mass in Pounds}
}
\value{
The mass in Kilograms
}
\description{
Convert mass in Pounds to Kilograms
}
\examples{
mass1 <- P_to_K(150);
mass2 <- P_to_K( c(200, 155, 342) );
}
